+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 ### Mods Added
 
+- HexxyPlanes: 0.2.0+1.20.1
 - Yet Another Hex Addon: 0.1.0
 
 ### Mods Updated